Move axis_chart*_test to axis_chart/ directory
