+++ test/chart/base/axis_chart/axis_chart_data_test.dart
@@ -1,7 +1,7 @@
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import '../data_pool.dart';
+import '../../data_pool.dart';
 
 void main() {
   group('AxisChartData data equality check', () {
