+++ test/chart/base/axis_chart/axis_chart_extensions_test.dart
@@ -2,7 +2,7 @@ import 'package:fl_chart/fl_chart.dart';
 import 'package:fl_chart/src/chart/base/axis_chart/axis_chart_extensions.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../data_pool.dart';
+import '../../data_pool.dart';
 
 void main() {
   group('splitByNullSpots()', () {
